refactor(cli): :recycle: move input methods to the `io.input` module
