+++ rocrate_validator/cli/commands/validate.py
@@ -69,70 +69,6 @@ def validate_uri(ctx, param, value):
     return value
 
 
-def __get_single_char_win32__(console: Optional[Console] = None, end: str = "\n",
-                              message: Optional[str] = None,
-                              choices: Optional[list[str]] = None) -> str:
-    """
-    Get a single character from the console
-    """
-    import msvcrt
-
-    char = None
-    while char is None or (choices and char not in choices):
-        if console and message:
-            console.print(f"\n{message}", end="")
-        try:
-            char = msvcrt.getch().decode()
-        finally:
-            if console:
-                console.print(char, end=end if choices and char in choices else "")
-        if choices and char not in choices:
-            if console:
-                console.print(" [bold red]INVALID CHOICE[/bold red]", end=end)
-    return char
-
-
-def __get_single_char_unix__(console: Optional[Console] = None, end: str = "\n",
-                             message: Optional[str] = None,
-                             choices: Optional[list[str]] = None) -> str:
-    """
-    Get a single character from the console
-    """
-    import termios
-    import tty
-
-    fd = sys.stdin.fileno()
-    old_settings = termios.tcgetattr(fd)
-    char = None
-    while char is None or (choices and char not in choices):
-        if console and message:
-            console.print(f"\n{message}", end="")
-        try:
-            tty.setraw(sys.stdin.fileno())
-            char = sys.stdin.read(1)
-            if char == "\x03":
-                raise KeyboardInterrupt
-        finally:
-            termios.tcsetattr(fd, termios.TCSADRAIN, old_settings)
-            if console:
-                console.print(char, end=end if choices and char in choices else "")
-        if choices and char not in choices:
-            if console:
-                console.print(" [bold red]INVALID CHOICE[/bold red]", end=end)
-    return char
-
-
-def get_single_char(console: Optional[Console] = None, end: str = "\n",
-                    message: Optional[str] = None,
-                    choices: Optional[list[str]] = None) -> str:
-    """
-    Get a single character from the console
-    """
-    if sys.platform == "win32":
-        return __get_single_char_win32__(console, end, message, choices)
-    return __get_single_char_unix__(console, end, message, choices)
-
-
 @cli.command("validate")
 @click.argument("rocrate-uri", callback=validate_uri, default=".")
 @click.option(
@@ -522,461 +458,3 @@ def validate(ctx,
         sys.exit(0 if is_valid else 1)
     except Exception as e:
         handle_error(e, console)
-
-
-def multiple_choice(console: Console,
-                    choices: list[Profile]):
-    """
-    Display a multiple choice menu
-    """
-    # Build the prompt text
-    prompt_text = "Please select the profiles to validate the RO-Crate against (<SPACE> to select):"
-
-    # Get the selected option
-    question = [
-        {
-            "type": "checkbox",
-            "name": "profiles",
-            "message": prompt_text,
-            "choices": [Choice(i, f"{choices[i].identifier}: {choices[i].name}") for i in range(0, len(choices))]
-        }
-    ]
-    console.print("\n")
-    selected = prompt(question, style={"questionmark": "#ff9d00 bold",
-                                       "question": "bold",
-                                       "checkbox": "magenta",
-                                       "answer": "magenta"},
-                      style_override=False)
-    logger.debug("Selected profiles: %s", selected)
-    return selected["profiles"]
-
-
-class ProgressMonitor(Subscriber):
-
-    PROFILE_VALIDATION = "Profiles"
-    REQUIREMENT_VALIDATION = "Requirements"
-    REQUIREMENT_CHECK_VALIDATION = "Requirements Checks"
-
-    def __init__(self, settings: dict, layout: ValidationReportLayout, stats: dict):
-        self.__progress = Progress(
-            TextColumn("[progress.description]{task.description}"),
-            BarColumn(),
-            TextColumn("{task.completed}/{task.total}"),
-            TimeElapsedColumn(),
-            expand=True)
-        self._stats = stats
-        self.settings = settings
-        self.profile_validation = self.progress.add_task(
-            self.PROFILE_VALIDATION, total=len(stats.get("profiles")))
-        self.requirement_validation = self.progress.add_task(
-            self.REQUIREMENT_VALIDATION, total=stats.get("total_requirements"))
-        self.requirement_check_validation = self.progress.add_task(
-            self.REQUIREMENT_CHECK_VALIDATION, total=stats.get("total_checks"))
-        self.__layout = layout
-        self._target_validation_profile = None
-        super().__init__("ProgressMonitor")
-
-    def start(self):
-        self.progress.start()
-
-    def stop(self):
-        self.progress.stop()
-
-    @property
-    def target_validation_profile(self) -> Profile:
-        return self._target_validation_profile
-
-    @target_validation_profile.setter
-    def target_validation_profile(self, profile: Profile):
-        self._target_validation_profile = profile
-
-    @property
-    def layout(self) -> ValidationReportLayout:
-        return self.__layout
-
-    @property
-    def progress(self) -> Progress:
-        return self.__progress
-
-    def update(self, event: Event):
-        # logger.debug("Event: %s", event.event_type)
-        if event.event_type == EventType.PROFILE_VALIDATION_START:
-            logger.debug("Profile validation start: %s", event.profile.identifier)
-        elif event.event_type == EventType.REQUIREMENT_VALIDATION_START:
-            logger.debug("Requirement validation start")
-        elif event.event_type == EventType.REQUIREMENT_CHECK_VALIDATION_START:
-            logger.debug("Requirement check validation start")
-        elif event.event_type == EventType.REQUIREMENT_CHECK_VALIDATION_END:
-            target_profile = self.target_validation_profile
-            if not event.requirement_check.requirement.hidden and \
-                    (not event.requirement_check.overridden
-                     or target_profile.identifier == event.requirement_check.requirement.profile.identifier):
-                self.progress.update(task_id=self.requirement_check_validation, advance=1)
-                if event.validation_result is not None:
-                    if event.validation_result:
-                        self._stats["passed_checks"].append(event.requirement_check)
-                    else:
-                        self._stats["failed_checks"].append(event.requirement_check)
-                    self.layout.update(self._stats)
-                else:
-                    logger.debug("Requirement check validation result is None: %s",
-                                 event.requirement_check.identifier)
-            else:
-                logger.debug("Skipping requirement check validation: %s", event.requirement_check.identifier)
-        elif event.event_type == EventType.REQUIREMENT_VALIDATION_END:
-            if not event.requirement.hidden:
-                self.progress.update(task_id=self.requirement_validation, advance=1)
-                if event.validation_result:
-                    self._stats["passed_requirements"].append(event.requirement)
-                else:
-                    self._stats["failed_requirements"].append(event.requirement)
-                self.layout.update(self._stats)
-        elif event.event_type == EventType.PROFILE_VALIDATION_END:
-            self.progress.update(task_id=self.profile_validation, advance=1)
-        elif event.event_type == EventType.VALIDATION_END:
-            self.layout.set_overall_result(event.validation_result)
-
-
-class ValidationReportLayout(Layout):
-
-    def __init__(self, console: Console, validation_settings: dict,
-                 profile_stats: dict, result: ValidationResult, profile_autodetected: bool = False):
-        super().__init__()
-        self.console = console
-        self.validation_settings = validation_settings
-        self.profile_stats = profile_stats
-        self.profile_autodetected = profile_autodetected
-        self.result = result
-        self.__layout = None
-        self._validation_checks_progress = None
-        self.__progress_monitor = None
-        self.requirement_checks_container_layout = None
-        self.passed_checks = None
-        self.failed_checks = None
-        self.report_details_container = None
-
-    @property
-    def layout(self):
-        if not self.__layout:
-            self.__init_layout__()
-        return self.__layout
-
-    @property
-    def validation_checks_progress(self):
-        return self._validation_checks_progress
-
-    @property
-    def progress_monitor(self) -> ProgressMonitor:
-        if not self.__progress_monitor:
-            self.__progress_monitor = ProgressMonitor(self.validation_settings, self, self.profile_stats)
-        return self.__progress_monitor
-
-    def live(self, update_callable: callable) -> any:
-        assert update_callable, "Update callable must be provided"
-        # Start live rendering
-        result = None
-        with Live(self.layout, console=self.console, refresh_per_second=10, transient=False):
-            result = update_callable()
-        return result
-
-    def __init_layout__(self):
-
-        # Get the validation settings
-        settings = self.validation_settings
-
-        # Set the console height
-        self.console.height = 31
-
-        # Create the layout of the base info of the validation report
-        severity_color = get_severity_color(Severity.get(settings["requirement_severity"]))
-        base_info_layout = Layout(
-            Align(
-                f"\n[bold cyan]RO-Crate:[/bold cyan] [bold]{URI(settings['rocrate_uri']).uri}[/bold]"
-                "\n[bold cyan]Target Profile:[/bold cyan][bold magenta] "
-                f"{settings['profile_identifier']}[/bold magenta] "
-                f"{'[italic](autodetected)[/italic]' if self.profile_autodetected else ''}"
-                f"\n[bold cyan]Validation Severity:[/bold cyan] "
-                f"[bold {severity_color}]{settings['requirement_severity']}[/bold {severity_color}]",
-                style="white", align="left"),
-            name="Base Info", size=5)
-        #
-        self.passed_checks = Layout(name="PASSED")
-        self.failed_checks = Layout(name="FAILED")
-        # Create the layout of the requirement checks section
-        validated_checks_container = Layout(name="Requirement Checks Validated")
-        validated_checks_container.split_row(
-            self.passed_checks,
-            self.failed_checks
-        )
-
-        # Create the layout of the requirement checks section
-        self.requirement_checks_by_severity_container_layout = Layout(name="Requirement Checks Validation", size=5)
-        self.requirement_checks_by_severity_container_layout.split_row(
-            Layout(name="required"),
-            Layout(name="recommended"),
-            Layout(name="optional")
-        )
-
-        # Create the layout of the requirement checks section
-        requirement_checks_container_layout = Layout(name="Requirement Checks")
-        requirement_checks_container_layout.split_column(
-            self.requirement_checks_by_severity_container_layout,
-            validated_checks_container
-        )
-
-        # Create the layout of the validation checks progress
-        self._validation_checks_progress = Layout(
-            Panel(Align(self.progress_monitor.progress, align="center"),
-                  border_style="white", padding=(0, 1), title="Overall Progress"),
-            name="Validation Progress", size=5)
-
-        # Create the layout of the report container
-        report_container_layout = Layout(name="Report Container Layout")
-        report_container_layout.split_column(
-            base_info_layout,
-            Layout(Panel(requirement_checks_container_layout,
-                   title="[bold]Requirements Checks Validation[/bold]", border_style="white", padding=(1, 1))),
-            self._validation_checks_progress
-        )
-
-        # Update the layout with the profile stats
-        self.update(self.profile_stats)
-
-        # Create the main layout
-        self.checks_stats_layout = Layout(
-            Panel(report_container_layout, title="[bold]- Validation Report -[/bold]",
-                  border_style="cyan", title_align="center", padding=(1, 2)))
-
-        # Create the overall result layout
-        self.overall_result = Layout(
-            Padding(Rule("\n[italic][cyan]Validating ROCrate...[/cyan][/italic]"), (1, 1)), size=3)
-
-        group_layout = Layout()
-        group_layout.add_split(self.checks_stats_layout)
-        group_layout.add_split(self.overall_result)
-
-        self.__layout = Padding(group_layout, (1, 1))
-
-    def update(self, profile_stats: dict = None):
-        assert profile_stats, "Profile stats must be provided"
-        self.profile_stats = profile_stats
-        self.requirement_checks_by_severity_container_layout["required"].update(
-            Panel(
-                Align(
-                    str(profile_stats['check_count_by_severity'][Severity.REQUIRED]) if profile_stats else "0",
-                    align="center"
-                ),
-                padding=(1, 1),
-                title="Severity: REQUIRED",
-                title_align="center",
-                border_style="RED"
-            )
-        )
-        self.requirement_checks_by_severity_container_layout["recommended"].update(
-            Panel(
-                Align(
-                    str(profile_stats['check_count_by_severity'][Severity.RECOMMENDED]) if profile_stats else "0",
-                    align="center"
-                ),
-                padding=(1, 1),
-                title="Severity: RECOMMENDED",
-                title_align="center",
-                border_style="orange1"
-            )
-        )
-        self.requirement_checks_by_severity_container_layout["optional"].update(
-            Panel(
-                Align(
-                    str(profile_stats['check_count_by_severity'][Severity.OPTIONAL]) if profile_stats else "0",
-                    align="center"
-                ),
-                padding=(1, 1),
-                title="Severity: OPTIONAL",
-                title_align="center",
-                border_style="yellow"
-            )
-        )
-
-        self.passed_checks.update(
-            Panel(
-                Align(
-                    str(len(self.profile_stats["passed_checks"])),
-                    align="center"
-                ),
-                padding=(1, 1),
-                title="PASSED Checks",
-                title_align="center",
-                border_style="green"
-            )
-        )
-
-        self.failed_checks.update(
-            Panel(
-                Align(
-                    str(len(self.profile_stats["failed_checks"])),
-                    align="center"
-                ),
-                padding=(1, 1),
-                title="FAILED Checks",
-                title_align="center",
-                border_style="red"
-            )
-        )
-
-    def set_overall_result(self, result: ValidationResult):
-        assert result, "Validation result must be provided"
-        self.result = result
-        if result.passed():
-            self.overall_result.update(
-                Padding(Rule("[bold][[green]OK[/green]] RO-Crate is a [green]valid[/green] "
-                             f"[magenta]{result.context.target_profile.identifier}[/magenta] !!![/bold]\n\n",
-                             style="bold green"), (1, 1)))
-        else:
-            self.overall_result.update(
-                Padding(Rule("[bold][[red]FAILED[/red]] RO-Crate is [red]not[/red] a [red]valid[/red] "
-                             f"[magenta]{result.context.target_profile.identifier}[/magenta] !!![/bold]\n",
-                             style="bold red"), (1, 1)))
-
-    def show_validation_details(self, pager: Pager, enable_pager: bool = True):
-        """
-        Print the validation result
-        """
-        if not self.result:
-            raise ValueError("Validation result is not available")
-
-        # init references to the console, result and severity
-        console = self.console
-        result = self.result
-
-        logger.debug("Validation failed: %s", result.failed_requirements)
-
-        # Print validation details
-        with console.pager(pager=pager, styles=not console.no_color) if enable_pager else console:
-            # Print the list of failed requirements
-            console.print(
-                Padding("\n[bold]The following requirements have not meet: [/bold]", (0, 2)), style="white")
-            for requirement in sorted(result.failed_requirements, key=lambda x: x.identifier):
-                console.print(
-                    Align(f"\n[profile: [magenta bold]{requirement.profile.name}[/magenta bold]]", align="right")
-                )
-                console.print(
-                    Padding(
-                        f"[bold][cyan][u][ {requirement.identifier} ]: "
-                        f"{Markdown(requirement.name).markup}[/u][/cyan][/bold]", (0, 5)), style="white")
-                console.print(Padding(Markdown(requirement.description), (1, 6)))
-                console.print(Padding("[white bold u]  Failed checks  [/white bold u]\n",
-                                      (0, 8)), style="white bold")
-
-                for check in sorted(result.get_failed_checks_by_requirement(requirement),
-                                    key=lambda x: (-x.severity.value, x)):
-                    issue_color = get_severity_color(check.level.severity)
-                    console.print(
-                        Padding(
-                            f"[bold][{issue_color}][ {check.identifier.center(16)} ][/{issue_color}]  "
-                            f"[magenta]{check.name}[/magenta][/bold]:", (0, 7)),
-                        style="white bold")
-                    console.print(Padding(Markdown(check.description), (0, 0, 0, len(check.identifier) + 13)))
-                    console.print(Padding("[u] Detected issues [/u]", (0, 8)), style="white bold")
-                    for issue in sorted(result.get_issues_by_check(check),
-                                        key=lambda x: (-x.severity.value, x)):
-                        path = ""
-                        if issue.violatingProperty and issue.violatingPropertyValue:
-                            path = f" of [yellow]{issue.violatingProperty}[/yellow]"
-                        if issue.violatingPropertyValue:
-                            if issue.violatingProperty:
-                                path += "="
-                            path += f"\"[green]{issue.violatingPropertyValue}[/green]\" "  # keep the ending space
-                        if issue.violatingEntity:
-                            path = f"{path} on [cyan]<{issue.violatingEntity}>[/cyan]"
-                        console.print(
-                            Padding(f"- [[red]Violation[/red]{path}]: "
-                                    f"{Markdown(issue.message).markup}", (0, 9)))
-                    console.print("\n", style="white")
-
-
-def __compute_profile_stats__(validation_settings: dict):
-    """
-    Compute the statistics of the profile
-    """
-    # extract the validation settings
-    severity_validation = Severity.get(validation_settings.get("requirement_severity"))
-    profiles = services.get_profiles(validation_settings.get("profiles_path"),
-                                     extra_profiles_path=validation_settings.get("extra_profiles_path"),
-                                     severity=severity_validation)
-    profile = services.get_profile(validation_settings.get("profile_identifier"),
-                                   validation_settings.get("profiles_path"),
-                                   extra_profiles_path=validation_settings.get("extra_profiles_path"),
-                                   severity=severity_validation)
-    target_profile_identifier = profile.identifier
-    # initialize the profiles list
-    profiles = [profile]
-
-    # add inherited profiles if enabled
-    if validation_settings.get("enable_profile_inheritance"):
-        profiles.extend(profile.inherited_profiles)
-    logger.debug("Inherited profiles: %r", profile.inherited_profiles)
-
-    # Initialize the counters
-    total_requirements = 0
-    total_checks = 0
-    check_count_by_severity = {}
-
-    # Initialize the counters
-    for severity in (Severity.REQUIRED, Severity.RECOMMENDED, Severity.OPTIONAL):
-        check_count_by_severity[severity] = 0
-
-    checks = []
-
-    # Process the requirements and checks
-    processed_requirements = []
-    for profile in profiles:
-        for requirement in profile.requirements:
-            processed_requirements.append(requirement)
-            if requirement.hidden:
-                continue
-
-            requirement_checks_count = 0
-            for severity in (Severity.REQUIRED, Severity.RECOMMENDED, Severity.OPTIONAL):
-                logger.debug(
-                    f"Checking requirement: {requirement} severity: {severity} {severity < severity_validation}")
-                # skip requirements with lower severity
-                if severity < severity_validation:
-                    continue
-                # count the checks
-                requirement_checks = [_ for _ in requirement.get_checks_by_level(LevelCollection.get(severity.name))
-                                      if not _.overridden or
-                                      _.requirement.profile.identifier == target_profile_identifier]
-                num_checks = len(requirement_checks)
-                check_count_by_severity[severity] += num_checks
-                requirement_checks_count += num_checks
-                checks.extend(requirement_checks)
-
-            # count the requirements and checks
-            if requirement_checks_count == 0:
-                logger.debug(f"No checks for requirement: {requirement}")
-            else:
-                logger.debug(f"Requirement: {requirement} checks count: {requirement_checks_count}")
-                assert not requirement.hidden, "Hidden requirements should not be counted"
-                total_requirements += 1
-                total_checks += requirement_checks_count
-
-    # log processed requirements
-    logger.debug("Processed requirements %r: %r", len(processed_requirements), processed_requirements)
-
-    # Prepare the result
-    result = {
-        "profile": profile,
-        "profiles": profiles,
-        "severity": severity_validation,
-        "check_count_by_severity": check_count_by_severity,
-        "total_requirements": total_requirements,
-        "total_checks": total_checks,
-        "failed_requirements": [],
-        "failed_checks": [],
-        "passed_requirements": [],
-        "passed_checks": [],
-        "checks": checks
-
-    }
-    logger.debug(result)
-    return result
